"minor" optimization changes

+++ modules/EngineCore/source/Board.cpp
@@ -1,13 +1,14 @@
+#include "CellState.h"
 #include "Board.h"
 
 namespace EngineCore {
 	Board::Board() {
 		reset();
 	}
-	CellState Board::get_CellState(int _row, int _col) const {
+	CellState Board::get_CellState(uint16_t _row, uint16_t _col) const {
 		return m_cells[_row][_col];
 	}
-	bool Board::set_CellState(int _row, int _col, CellState _state) {
+	bool Board::set_CellState(uint16_t _row, uint16_t _col, CellState _state) {
 		if (m_cells[_row][_col] == CellState::EMPTY) {
 			m_cells[_row][_col] = _state;
 			return true;
